Election results according to Central Election Board

A dataset with the data contained on the PDF documents provided since 2004 by Central Election Board (CEB) in Spain. This dataset is used in the unit tests since the data provided by the Ministry of the Interior are not always corrected by the CEB data (which are always at the province level, not below). The tests check that such discrepancies do not exceed 0.5% of the votes.

Usage

data(CEB_results)

Format

A tibble with 15 columns

id_elec: id of elections (type of elections + date).

cod_elec: code of elections.

date: date of elections.

cod_INE_ccaa, cod_MIR_ccaa, ccaa: codes and names for ccaa.

cod_INE_prov, id_INE_prov, prov: codes and names for provinces.

census_counting_prov: census for each election and province.

total_ballots: total number of ballots, as a sum of valid_ballots and invalid_ballots.

valid_ballots: total number of valid ballots, as a sum of party_ballots and blank_ballots.

party_ballots: total number of ballots to candidacies.

blank_ballots: total number of blank ballots.

invalid_ballots: total number of invalid ballots.

Source

Data was scraped from the PDF documents provided since 2004 by Central Election Board (CEB) in Spain.

Author

Javier Alvarez-Liebana.

Examples

# Load data
data(CEB_results)