Import candidacies data

Import and preprocess candidacies data. This function supports both single values and vector inputs for fetching and combining data for multiple elections at once.

Usage

import_candidacies_data(
  type_elec,
  year = 2019,
  month = 4,
  date = NULL,
  repo_url = "https://github.com/dadosdelaplace/pollspain-data/blob/main",
  file_ext = ".rda",
  short_version = TRUE,
  verbose = TRUE
)

Arguments

type_elec

Type elections for which data is available. It should be one of the following values: "referendum", "congress", "senate", "local", "cabildo" (Canarian council) or "EU".

year

A vector or single value representing the years of the elections to be considered.

month

A vector or single value representing the months of the elections to be considered. Please, check (refer to dates_elections_spain) that elections of the specified type are available for the provided year and month.

date

A vector or single value representing the dates of the elections to be considered. If date was provided, it should be in format %Y-%m-%d (e.g., '2000-01-01'). Defaults to NULL. If no date was provided, month and year should be both numerical variables.

repo_url

A string with the url in which the raw data can be found. Defaults to https://github.com/dadosdelaplace/pollspain-data/blob/main.

file_ext

A string with the file's extension. Defaults to ".rda".

short_version

flag to indicate whether it should be returned a short version of the data (just key variables) or not. Defaults to TRUE.

verbose

Flag to indicate whether detailed messages should be printed during execution. Defaults to TRUE.

Value

A tibble with candidacies data at poll station level including the following variables:

id_elec

election's id constructed from the election code cod_elec and date date_elec.

cod_elec

code representing the type of election: "01" (referendum), "02" (congress), "03" (senate), "04" (local elections), "06" (cabildo - Canarian council - elections), "07" (European Parliament elections). Variable available only for long version.

type_elec

type of election.

date_elec

date of the election.

id_INE_poll_station

poll station's id constructed from the ccaa-prov-municipality and poll station codes.

id_INE_mun

municipality ID constructed from the ccaa-prov-mun codes provided by INE.

cod_INE_ccaa, ccaa

codes and names for regions (ccaa) to which the municipalities belong. Codes available only for long version.

cod_INE_prov, prov

codes and names for the provinces to which the municipalities belong. Codes available only for long version.

cod_INE_mun, mun

code, and name for municipalities. Codes available only for long version.

cod_mun_district, cod_sec, cod_poll_station

codes for the municipal district, census tract and poll station. Codes available only for long version.

id_candidacies

id for candidacies (at province level).

id_candidacies_ccaa, id_candidacies_nat

id for candidacies (at region - ccaa - and national level). Id's available only for long version.

abbrev_candidacies, name_candidacies

acronym and full name of the candidacies.

ballots

number of ballots obtained for each candidacy at each poll station.

Details

This function fetches candidates data for the specified elections by downloading the corresponding .rda files from GitHub https://github.com/dadosdelaplace/pollspain-data (directly downloaded from MIR website) and processing them into a tidy format. It automatically handles the download, loading, and merging of data across multiple election periods as specified by the user.

Author

Javier Álvarez-Liébana, David Pereiro Pol, Mafalda González González, Irene Bosque Gala and Mikaela De Smedt.

Examples

## Correct examples

# Congress elections in April 2019
# Fetch candidacies data for the congress elections in April 2019
candidacies_data1 <- import_candidacies_data("congress", 2019, 4)
#> 🔎 Check if parameters are allowed...
#> Warning: tz(): Don't know how to compute timezone for object of class numeric; returning "UTC".
#> Error in import_candidacies_data("congress", 2019, 4): 😵 No congress elections are available. Please, be sure that arguments and dates are right

# Fetch candidacies data for the congress and senate elections
# in April 2019
candidacies_data2 <-
  import_candidacies_data(c("congress", "senate"), 2019, 4)
#> 🔎 Check if parameters are allowed...
#> Warning: tz(): Don't know how to compute timezone for object of class numeric; returning "UTC".
#> Error in import_candidacies_data(c("congress", "senate"), 2019, 4): 😵 No congress elections are available. Please, be sure that arguments and dates are right😵 No senate elections are available. Please, be sure that arguments and dates are right

# Example usage to combine data from different elections into
# one table
# Fetch candidacies data for congress elections in Nov 2019
# and June 2016
combined_candidacies_data <-
  import_candidacies_data(c("congress", "congress"),
                          c(2019, 2016), c(11, 6))
#> 🔎 Check if parameters are allowed...
#> Warning: tz(): Don't know how to compute timezone for object of class numeric; returning "UTC".
#> Error in if (is.na(date)) {    stop(red("😵 If date was provided, `date` should be in format '2000-01-01' (%Y-%m-%d)"))}: the condition has length > 1

# Example usage providing date instead of month and year
date_example <-
  import_candidacies_data("congress", date = "2016-06-26")
#> 🔎 Check if parameters are allowed...
#> 📦 Import candidacies data at poll station level from ...
#> - https://github.com/dadosdelaplace/pollspain-data/blob/main/data/02-congress/02201606/raw_candidacies_poll_congress_2016_06.rda?raw=true
#> 
#> ⏳ Please wait, the volume of data downloaded and the internet connection may take a few seconds
#> ⚠️ A short version was asked. If you require all variables, please run with `short_version = FALSE'

# ----
# Incorrect examples
# ----

if (FALSE) { # \dontrun{
# Wrong examples

# Invalid election type: "national" is not a valid election type
import_candidacies_data("national", 2019, 4)

# Invalid election: no congress elections are available in 5-2019.
# Please check dataset dates_elections_spain
import_candidacies_data("congress", 2019, 5)

# Invalid date format: date should be in %Y-%m-%d format
import_candidacies_data("congress", date = "26-06-2016")

} # }