Import municipal census data

Import municipal census data for one or more elections at the municipal level. This function downloads and processes raw municipal data files for specified elections.

Usage

import_mun_census_data(
  type_elec,
  year = 2019,
  month = 4,
  date = NULL,
  repo_url = "https://github.com/dadosdelaplace/pollspain-data/blob/main",
  file_ext = ".rda",
  verbose = TRUE
)

Source

Some definitions of variables were extracted from https://www.ige.gal.

Arguments

type_elec

Type elections for which data is available. It should be one of the following values: "referendum", "congress", "senate", "local", "cabildo" (Canarian council) or "EU".

year

A vector or single value representing the years of the elections to be considered.

month

A vector or single value representing the months of the elections to be considered. Please, check (refer to dates_elections_spain) that elections of the specified type are available for the provided year and month.

date

A vector or single value representing the dates of the elections to be considered. If date was provided, it should be in format %Y-%m-%d (e.g., '2000-01-01'). Defaults to NULL. If no date was provided, month and year should be both numerical variables.

repo_url

A string with the url in which the raw data can be found. Defaults to https://github.com/dadosdelaplace/pollspain-data/blob/main.

file_ext

A string with the file's extension. Defaults to ".rda".

verbose

Flag to indicate whether detailed messages should be printed during execution. Defaults to TRUE.

Value

A tibble with rows corresponding to municipalities for each election, including the following variables:

id_elec

election's id constructed from the election code cod_elec and date date_elec.

cod_elec

code representing the type of election: "01" (referendum), "02" (congress), "03" (senate), "04" (local elections), "06" (cabildo - Canarian council - elections), "07" (European Parliament elections).

type_elec

type of election.

date_elec

date of the election.

id_INE_mun

municipality ID constructed from the ccaa-prov-mun codes provided by INE.

cod_INE_ccaa, ccaa

codes and names for regions (ccaa) to which the municipalities belong.

cod_INE_prov, prov

codes and names for the provinces to which the municipalities belong.

cod_INE_mun, mun

code, and name for municipalities.

cod_mun_jud_district, cod_mun_prov_council

codes for the judicial district and provincial council.

n_poll_stations

number of polling stations in each municipality.

pop_res_mun

population census of residents (CER + CERA) at municipality level.

census_INE_mun

population eligible to vote at municipality level.

census_counting_mun

population eligible to vote after claims at municipality level.

census_CERE_mun

census of foreign nationals, relevant only for EU elections at municipality level.

Details

This function fetches municipal-level data for the specified elections by downloading the corresponding .rda files from GitHub https://github.com/dadosdelaplace/pollspain-data (directly downloaded from MIR website) and processing them into a tidy format. It automatically handles the download, loading, and merging of data across multiple election periods as specified by the user.

Author

Javier Álvarez-Liébana, David Pereiro Pol, Mafalda González González, Irene Bosque Gala and Mikaela De Smedt.

Examples

## Correct examples

# Congress elections in April 2019
# Fetch municipal census data for the congress elections in
# April 2019
mun_census_data1 <- import_mun_census_data("congress", 2019, 4)
#> 🔎 Check if parameters are allowed...
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Senate elections in April 2019
# Fetch municipal census data for the senate elections in
# April 2019
mun_census_data2 <- import_mun_census_data("senate", 2019, 4)
#> 🔎 Check if parameters are allowed...
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Example usage to combine data from different elections into
# one table
# Fetch municipal census data for congress elections in Nov 2019
# and June 2016
combined_mun_census_data <-
  import_mun_census_data(c("congress", "congress"),
                         c(2019, 2016), c(11, 6))
#> 🔎 Check if parameters are allowed...
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Example usage providing date instead of month and year
date_example <-
  import_mun_census_data("congress", date = "2016-06-26")
#> 🔎 Check if parameters are allowed...
#> 📦 Import census mun data from ...
#> - https://github.com/dadosdelaplace/pollspain-data/blob/main/data/02-congress/02201606/raw_mun_data_congress_2016_06.rda?raw=true
#> 

# ----
# Incorrect examples
# ----
if (FALSE) {
# Wrong examples

# Invalid election type: "national" is not a valid election type
import_mun_census_data("national", 2019, 4)

# Invalid election: no congress elections are available in 5-2019.
# Please check dataset dates_elections_spain
import_mun_census_data("congress", 2019, 5)

# Invalid date format: date should be in %Y-%m-%d format
import_mun_census_data("congress", date = "26-06-2016")
}