Recoding Spanish political party or candidacies names and preprocessing of acronyms

This function standardizes and recodes the names and abbreviations of Spanish political party names in a given dataset. It handles various text transformations, such as removing special characters, trimming whitespace, and converting characters to ASCII. The function also recodes certain party names and abbreviations to standardized formats based on predefined patterns.

Arguments

parties_data

A dataset (preferably a tibble) containing political party data. Function will expect a dataset with 2 columns for party acronyms and full names.

col_acronym

A character string specifying the column name for party acronyms in parties_data. Defaults to "abbrev_candidacies".

col_full_name

A character string specifying the column name for party full names in parties_data. Defaults to "name_candidacies".

Value

A tibble with standardized and recoded political party names and acronyms. The output datasets includes cleaned versions of the original columns and recoded values based on predefined patterns.

Details

The function first renames the specified columns in the input dataset to a standard format. It then performs a series of text cleaning operations, such as removing special characters, replacing certain symbols with hyphens, and converting accented characters to their ASCII equivalents. After cleaning the text, the function applies a series of pattern-matching rules to recode specific party names and abbreviations to standardized versions.

Author

Javier Álvarez-Liébana, David Pereiro-Pol, Mafalda González-González and Irene Bosque-Gala.

Examples

if (FALSE) {
# Assuming `party_data` is a dataframe with columns
# "abbrev" and "full_name"
standardized_data <-
  recod_parties(party_data, col_name_abbrev = "abbrev",
                col_name_candidacies = "full_name")
}