Summaries of the electoral and candidacies ballots data for a given aggregation level (ccaa, prov, etc)

pending...

Usage

summary_election_data(
  type_elec,
  year = 2019,
  month = 4,
  date = NULL,
  election_data = NULL,
  ballots_data = NULL,
  col_id_elec = "id_elec",
  col_id_poll_station = "id_INE_poll_station",
  prec_round = 3,
  short_version = TRUE,
  repo_url = "https://github.com/dadosdelaplace/pollspain-data/blob/main",
  file_ext = ".rda",
  level = "all",
  by_parties = FALSE,
  filter_porc_ballots = NA,
  cols_mun_var = c("pop_res_mun", "census_counting_mun"),
  col_id_candidacies = "id_candidacies",
  cols_names_candidacies = c("abbrev_candidacies", "name_candidacies"),
  verbose = TRUE
)

Arguments

type_elec

Type elections for which data is available. It should be one of the following values: "referendum", "congress", "senate", "local", "cabildo" (Canarian council) or "EU".

year

A vector or single value representing the years of the elections to be considered.

month

A vector or single value representing the months of the elections to be considered. Please, check (refer to dates_elections_spain) that elections of the specified type are available for the provided year and month.

date

A vector or single value representing the dates of the elections to be considered. If date was provided, it should be in format %Y-%m-%d (e.g., '2000-01-01'). Defaults to NULL. If no date was provided, month and year should be both numerical variables.

election_data

A database containing general election data already provided (by other functions or by the user). Database should contain a col_id_elec column. Defaults to NULL.

ballots_data

A database containing ballots data for the different candidacies, already provided (by other functions or by the user). Database should contain a col_id_poll_station column. Defaults to NULL.

col_id_elec

Column name in election_data to uniquely identify the elections. Defaults to "id_elec".

col_id_poll_station

Column name in ballots_data to uniquely identify poll stations. Defaults to "id_INE_poll_station".

prec_round

Rounding accuracy. Defaults to prec_round = 3.

short_version

Flag to indicate whether it should be returned a short version of the data (just key variables) or not. Defaults to TRUE.

repo_url

A string with the url in which the raw data can be found. Defaults to https://github.com/dadosdelaplace/pollspain-data/blob/main.

file_ext

A string with the file's extension. Defaults to ".rda".

level

A string providing the level of aggregation at which the data is to be provided. The allowe values are the following: 'all', 'ccaa', 'prov', 'mun', 'mun_district', 'sec' or 'poll_station'. Defaults to "all".

by_parties

A flag indicates whether user wants a summary by candidacies/parties or just global results at given level. Defaults to FALSE.

filter_porc_ballots

A numerical argument representing the vote percentage threshold (out of 100) that the user wants to use to filter the parties (as long as by_parties = TRUE). Defaults to NA.

cols_mun_var

A vector of variable names that, in their raw version, are only available at the municipal level (or higher). Defaults to c("pop_res_mun", "census_counting_mun").

col_id_candidacies

A string indicating the name of the column that uniquely identifies the candidacies. Defaults to "id_candidacies".

cols_names_candidacies

A string indicating the name of the column containg the names and/or acronyms of candidacies. Defaults to c("abbrev_candidacies", "name_candidacies").

verbose

Flag to indicate whether detailed messages should be printed during execution. Defaults to TRUE.

Value

A tibble with rows corresponding to ...pending

Details

This function ... pending

Author

Javier Álvarez-Liébana and David Pereiro-Pol.

Examples

## Correct examples

# Summary election data at national level (general data
# without candidacies ballots)
summary_data_all <- summary_election_data("congress", 2019, 4)
#> 🔎 Check if parameters are allowed...
#> 📦 Import data at poll station level ...
#> ⏳ Please wait, the volume of data downloaded and the internet connection may take a few seconds
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Summary election data at national level, aggregating the
# candidacies ballots
summary_data_all_parties <-
  summary_election_data("congress", 2019, 4, by_parties = TRUE)
#> 🔎 Check if parameters are allowed...
#> 📦 Import data at poll station level ...
#> ⏳ Please wait, the volume of data downloaded and the internet connection may take a few seconds
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Summary election data at ccaa level (general data
# without candidacies ballots)
summary_data_ccaa <-
  summary_election_data("congress", 2019, 4, level = "ccaa")
#> 🔎 Check if parameters are allowed...
#> 📦 Import data at poll station level ...
#> ⏳ Please wait, the volume of data downloaded and the internet connection may take a few seconds
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

# Summary election data at ccaa level, aggregating the
# candidacies ballots
summary_data_ccaa_parties <-
  summary_election_data("congress", 2019, 4, level = "ccaa",
                        by_parties = TRUE)
#> 🔎 Check if parameters are allowed...
#> 📦 Import data at poll station level ...
#> ⏳ Please wait, the volume of data downloaded and the internet connection may take a few seconds
#> Error in as.POSIXlt.numeric(x, tz = tz(x)): 'origin' must be supplied

if (FALSE) {
# ----
# Incorrect examples
# ----

# Wrong examples

}