Summaries of the electoral and candidacies ballots data for a given aggregation level (ccaa, prov, etc)

pending Import, preprocess and aggregate election data at the same time for a given election and aggregation level. This function also lets remove parties below a given vote share threshold.

Usage

summary_election_data(
  type_elec,
  year = NULL,
  date = NULL,
  election_data = NULL,
  ballots_data = NULL,
  candidacies_data = NULL,
  col_id_elec = "id_elec",
  col_id_poll_station = "id_INE_poll_station",
  prec_round = 3,
  short_version = TRUE,
  CERA_remove = FALSE,
  level = "all",
  by_parties = TRUE,
  filter_porc_ballots = NA,
  filter_candidacies = NA,
  cols_mun_var = c("pop_res_mun", "census_counting_mun"),
  col_id_candidacies = c(id_prov = "id_candidacies", id_nat = "id_candidacies_nat"),
  col_abbrev_candidacies = "abbrev_candidacies",
  verbose = TRUE
)

Arguments

type_elec

Type elections for which data is available. It should be one of the following values: "referendum", "congress", "senate", "local", "cabildo" (Canarian council) or "EU".

year

A vector or single value representing the years of the elections to be considered. Please, check in dates_elections_spain that elections of the specified type are available for the provided year.

date

A vector or single value representing the dates of the elections to be considered. If date was provided, it should be in format %Y-%m-%d (e.g., '2000-01-01'). Defaults to NULL. If no date was provided, year should be provided as numerical variable. Please, check in dates_elections_spain that elections of the specified type are available.

election_data

A database containing general election data already provided (by other functions or by the user). Database should contain a col_id_elec column. Defaults to NULL.

ballots_data

A database containing ballots data for the different candidacies, already provided (by other functions or by the user). Database should contain a col_id_poll_station column. Defaults to NULL.

col_id_elec

Column name in election_data to uniquely identify the elections. Defaults to "id_elec".

col_id_poll_station

Column name in ballots_data to uniquely identify poll stations. Defaults to "id_INE_poll_station".

prec_round

Rounding accuracy. Defaults to prec_round = 3.

short_version

Flag to indicate whether it should be returned a short version of the data (just key variables) or not. Defaults to TRUE.

CERA_remove

pending

level

A string providing the level of aggregation at which the data is to be provided. The allowed values are the following: 'all', 'ccaa', 'prov', 'mun', 'mun_district', 'sec' or 'poll_station'. Defaults to "all".

by_parties

A flag indicates whether user wants a summary by candidacies/parties or just global results at given level. Defaults to TRUE.

filter_porc_ballots

A numerical argument representing the vote percentage threshold (out of 100) that the user wants to use to filter the parties (as long as by_parties = TRUE). Defaults to NA.

filter_candidacies

A string of characters ... pending (as long as by_parties = TRUE). Defaults to NA.

cols_mun_var

A vector of variable names that, in their raw version, are only available at the municipal level (or higher). Defaults to c("pop_res_mun", "census_counting_mun").

col_id_candidacies

A named vector (with names "id_prov" and "id_nat") in which user provides the column names for the candidacies id at province level and national level. Defaults to c("id_prov" = "id_candidacies", "id_nat" = "id_candidacies_nat").

col_abbrev_candidacies

pending...

verbose

Flag to indicate whether detailed messages should be printed during execution. Defaults to TRUE.

Value

A tibble with rows corresponding to the level of aggregation for each election, including the following variables:

id_elec

election's id constructed from the election code cod_elec and date date_elec.

cod_elec

code representing the type of election: "01" (referendum), "02" (congress), "03" (senate), "04" (local elections), "06" (cabildo - Canarian council - elections), "07" (European Parliament elections).

type_elec

type of election.

date_elec

date of the election.

id_INE_poll_station

poll station's id constructed from the ccaa-prov-municipality and poll station codes.

id_INE_mun

municipality ID constructed from the ccaa-prov-mun codes provided by INE.

cod_INE_ccaa, ccaa

codes and names for regions (ccaa) to which the municipalities belong.

cod_INE_prov, prov

codes and names for the provinces to which the municipalities belong.

cod_INE_mun, mun

code, and name for municipalities.

cod_mun_district, cod_sec, cod_poll_station

codes for the municipal district, census tract and poll station.

ballots_1, turnout_1

number of total ballots and turnout percentage in the first round.

ballots_2, turnout_2

number of total ballots and turnout percentage in the second round (if applicable).

blank_ballots, invalid_ballots

blank and invalid ballots.

party_ballots, valid_ballots, total_ballots

ballots to candidacies/parties, valid ballots (sum of blank_ballots and party_ballots) and total ballots (sum of valid_ballots and invalid_ballots).

turnout

final turnout percentage.

porc_valid, porc_invalid, porc_parties, porc_blank

perc (%) values of valid_ballots, invalid_ballots, party_ballots and blank_ballots.

n_poll_stations

number of polling stations.

pop_res_mun

population census of residents (CER + CERA) at municipality level.

census_counting_mun

population eligible to vote after claims at municipality level.

id_candidacies

id for candidacies (at province level).

id_candidacies_ccaa, id_candidacies_nat

id for candidacies (at region - ccaa - and national level).

abbrev_candidacies, name_candidacies

acronym and full name of the candidacies.

ballots

number of ballots obtained for each candidacy at each level section.

porc_candidacies_parties

percentage of ballots obtained for each candidacy over the total number of party ballots of each level section.

porc_candidacies_valid

percentage of ballots obtained for each candidacy over the total number of valid ballots of each level section.

porc_candidacies_census

percentage of ballots obtained for each candidacy over the census of the corresponding aggregation level.

Details

A tibble with rows corresponding to the level of aggregation for each election, including the following variables:

Author

Javier Alvarez-Liebana and David Pereiro-Pol.

Examples

## Correct examples

# Summary 2023 and 2016 election data at national level, without
# candidacies ballots, in a short version
summary_all <-
  summary_election_data(type_elec = "congress", year = 2023,
                        date = "2016-06-26", short_version = TRUE,
                        by_parties = FALSE)
#>    ... please wait, the volume of data to be aggregated may take a few seconds
#> ... Check if parameters are allowed...
#> [x] Import poll station data ...
#>    ... congress elections on 2023-07-24
#>    ... congress elections on 2016-06-26
#> 
#> [x] Import candidacies and ballots data (at poll station level) ...
#>    ... please wait, volume of data downloaded and internet connection may take a few seconds
#> [x] Aggregate data at national level ...
#>    ... check if aggregation parameters are allowed...
#> [x] Join information sources and last summaries ...
#> 

# Summary 2023 and 2016 election data at national level,
# aggregating the candidacies ballots, in a long version
summary_all <-
  summary_election_data(type_elec = "congress", year = 2023,
                        date = "2016-06-26", short_version = FALSE)
#>    ... please wait, the volume of data to be aggregated may take a few seconds
#> ... Check if parameters are allowed...
#> [x] Import poll station data ...
#>    ... congress elections on 2023-07-24
#>    ... congress elections on 2016-06-26
#> 
#> [x] Import candidacies and ballots data (at poll station level) ...
#>    ... please wait, volume of data downloaded and internet connection may take a few seconds
#> [x] Aggregate data at national level ...
#>    ... check if aggregation parameters are allowed...
#> [x] Join information sources and last summaries ...
#> 

# Summary 2023 election data at ccaa level, aggregating the
# candidacies ballots, in a short version
summary_ccaa <-
  summary_election_data(type_elec = "congress", year = 2023,
                        date = "2016-06-26", level = "ccaa",
                        short_version = FALSE)
#>    ... please wait, the volume of data to be aggregated may take a few seconds
#> ... Check if parameters are allowed...
#> [x] Import poll station data ...
#>    ... congress elections on 2023-07-24
#>    ... congress elections on 2016-06-26
#> 
#> [x] Import candidacies and ballots data (at poll station level) ...
#>    ... please wait, volume of data downloaded and internet connection may take a few seconds
#> [x] Aggregate data at ccaa level ...
#>    ... check if aggregation parameters are allowed...
#> [x] Join information sources and last summaries ...
#> 

# Summary 2023 election data at prov level, aggregating the
# candidacies ballots, in a short version, and filtering ballots
# (percentage between 0 and 100)
summary_prov <-
  summary_election_data(type_elec = "congress", year = 2023,
                        date = "2016-06-26", level = "prov",
                        short_version = FALSE,
                        filter_porc_ballots = 15)
#>    ... please wait, the volume of data to be aggregated may take a few seconds
#> ... Check if parameters are allowed...
#> [x] Import poll station data ...
#>    ... congress elections on 2023-07-24
#>    ... congress elections on 2016-06-26
#> 
#> [x] Import candidacies and ballots data (at poll station level) ...
#>    ... please wait, volume of data downloaded and internet connection may take a few seconds
#> [x] Aggregate data at prov level ...
#>    ... check if aggregation parameters are allowed...
#> [x] Join information sources and last summaries ...
#> 

if (FALSE) {
# ----
# Incorrect examples
# ----

# Wrong examples

# Invalid aggregation level
summary_election_data("congress", 2019, level = "district")

# filter_porc_ballots outside range 0 from 100
summary_election_data("congress", 2019,
                      filter_porc_ballots = 150)

# filter_porc_ballots supplied while by_parties = FALSE
summary_election_data("congress", 2019,
                      by_parties = FALSE,
                      filter_porc_ballots = 5)

# Wrong election type
summary_election_data("national", 2019)
}